Order: Solved bug on total counting on discounted Condition prices

+++ src/main/java/com/unisa/store/tsw_project/controller/user/TheOrderServlet.java
@@ -283,7 +283,7 @@ public class TheOrderServlet extends HttpServlet {
                 }
 
                 //Apply any Condition Discount
-                discountedPrice = discountedPrice.multiply(BigDecimal.valueOf((1 - cond.getCondition().getDiscount()/100)));
+                discountedPrice = discountedPrice.multiply(BigDecimal.valueOf((1 - (double) cond.getCondition().getDiscount()/100.0)));
             }
 
             /* Apply Any DiscountCodeValue */
